component for person link was created
