+++ src/components/PeopleTable.tsx
@@ -1,32 +1,26 @@
 import { useContext } from 'react';
-import {
-  Link,
-  useLocation,
-  useResolvedPath,
-} from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import classNames from 'classnames';
 
 import { Person } from '../types/Person';
 import { getSearchWith } from '../utils/searchHelper';
 import { SearchParamsContext } from './SearchParamsContext';
+import { PersonLink } from './PersonLink';
 
 const sortTypes = ['Name', 'Sex', 'Born', 'Died'];
 
 type Props = {
   people: Person[],
-  selectedPersonSlug: string | undefined,
 };
 
 export const PeopleTable: React.FC<Props> = ({
   people,
-  selectedPersonSlug,
 }) => {
   let visiblePeople = [...people];
 
-  const { searchParams } = useContext(SearchParamsContext);
+  const { slug: selectedPersonSlug } = useParams();
 
-  const resolvedPath = useResolvedPath('../').pathname;
-  const location = useLocation();
+  const { searchParams } = useContext(SearchParamsContext);
 
   const sort = searchParams.get('sort');
   const order = searchParams.get('order');
@@ -81,60 +75,59 @@ export const PeopleTable: React.FC<Props> = ({
   }
 
   return (
-    <table
-      data-cy="peopleTable"
-      className="table is-striped is-hoverable is-narrow is-fullwidth"
-    >
-      <thead>
-        <tr>
-          {sortTypes.map(name => {
-            const nameLowercased = name.toLowerCase();
-            const isCurrentSortType = sort === nameLowercased;
+    visiblePeople.length
+      ? (
+        <table
+          data-cy="peopleTable"
+          className="table is-striped is-hoverable is-narrow is-fullwidth"
+        >
+          <thead>
+            <tr>
+              {sortTypes.map(name => {
+                const nameLowercased = name.toLowerCase();
+                const isCurrentSortType = sort === nameLowercased;
 
-            return (
-              <th key={name}>
-                <span className="is-flex is-flex-wrap-nowrap">
-                  {name}
-                  <Link
-                    to={{
-                      search: getSearchWith(
-                        searchParams,
-                        {
-                          sort: isCurrentSortType && order
-                            ? null : nameLowercased,
-                          order: isCurrentSortType && !order
-                            ? 'desc' : null,
-                        },
-                      ),
-                    }}
-                  >
-                    <span className="icon">
-                      <i
-                        className={classNames(
-                          'fas',
-                          { 'fa-sort': !isCurrentSortType },
-                          { 'fa-sort-up': isCurrentSortType && !order },
-                          { 'fa-sort-down': isCurrentSortType && order },
-                        )}
-                      />
+                return (
+                  <th key={name}>
+                    <span className="is-flex is-flex-wrap-nowrap">
+                      {name}
+                      <Link
+                        to={{
+                          search: getSearchWith(
+                            searchParams,
+                            {
+                              sort: isCurrentSortType && order
+                                ? null : nameLowercased,
+                              order: isCurrentSortType && !order
+                                ? 'desc' : null,
+                            },
+                          ),
+                        }}
+                      >
+                        <span className="icon">
+                          <i
+                            className={classNames(
+                              'fas',
+                              { 'fa-sort': !isCurrentSortType },
+                              { 'fa-sort-up': isCurrentSortType && !order },
+                              { 'fa-sort-down': isCurrentSortType && order },
+                            )}
+                          />
+                        </span>
+                      </Link>
                     </span>
-                  </Link>
-                </span>
-              </th>
-            );
-          })}
+                  </th>
+                );
+              })}
 
-          <th>Mother</th>
-          <th>Father</th>
-        </tr>
-      </thead>
+              <th>Mother</th>
+              <th>Father</th>
+            </tr>
+          </thead>
 
-      <tbody>
-        {visiblePeople.length
-          ? (
-            visiblePeople.map(person => {
+          <tbody>
+            {visiblePeople.map(person => {
               const {
-                name,
                 sex,
                 born,
                 died,
@@ -169,17 +162,7 @@ export const PeopleTable: React.FC<Props> = ({
                   )}
                 >
                   <td>
-                    <Link
-                      to={{
-                        pathname: resolvedPath + slug,
-                        search: location.search,
-                      }}
-                      className={classNames(
-                        { 'has-text-danger': sex === 'f' },
-                      )}
-                    >
-                      {name}
-                    </Link>
+                    <PersonLink person={person} />
                   </td>
 
                   <td>
@@ -197,15 +180,7 @@ export const PeopleTable: React.FC<Props> = ({
                   <td>
                     {mother
                       ? (
-                        <Link
-                          to={{
-                            pathname: resolvedPath + mother.slug,
-                            search: location.search,
-                          }}
-                          className="has-text-danger"
-                        >
-                          {motherName}
-                        </Link>
+                        <PersonLink person={mother} />
                       )
                       : (
                         motherName || '-'
@@ -215,14 +190,7 @@ export const PeopleTable: React.FC<Props> = ({
                   <td>
                     {father
                       ? (
-                        <Link
-                          to={{
-                            pathname: resolvedPath + father.slug,
-                            search: location.search,
-                          }}
-                        >
-                          {fatherName}
-                        </Link>
+                        <PersonLink person={father} />
                       )
                       : (
                         fatherName || '-'
@@ -230,14 +198,14 @@ export const PeopleTable: React.FC<Props> = ({
                   </td>
                 </tr>
               );
-            })
-          )
-          : (
-            <p data-cy="noPeopleMessage">
-              There are no people on the server
-            </p>
-          )}
-      </tbody>
-    </table>
+            })}
+          </tbody>
+        </table>
+      )
+      : (
+        <p>
+          There are no people matching the current search criteria
+        </p>
+      )
   );
 };
